# httpserver/publicapi/api.gen.go
//go:build go1.22

// Package publicapi provides primitives to interact with the openapi HTTP API.
//
// Code generated by github.com/oapi-codegen/oapi-codegen/v2 version v2.4.1 DO NOT EDIT.
package publicapi

import (
	"fmt"
	"net/http"

	"github.com/oapi-codegen/runtime"
)

// Address 20-byte Ethereum address as a `0x`-prefixed hex string.
type Address = string

// AddressBalance defines model for AddressBalance.
type AddressBalance struct {
	// Address 20-byte Ethereum address as a `0x`-prefixed hex string.
	Address Address `json:"address"`

	// BalanceHex Native balance as a `0x`-prefixed hex quantity.
	BalanceHex string `json:"balanceHex"`

	// BalanceWei Native balance in wei as a base-10 string.
	BalanceWei string `json:"balanceWei"`

	// Block Requested block tag (`latest` or decimal block number).
	Block string `json:"block"`
}

// Block defines model for Block.
type Block struct {
	// BlockHash Block hash.
	BlockHash string `json:"blockHash"`

	// BlockNumber Block number as a decimal string.
	BlockNumber string `json:"blockNumber"`

	// Nonce Block nonce (`0x`-prefixed when present).
	Nonce string `json:"nonce"`

	// Timestamp Block timestamp in milliseconds since Unix epoch.
	Timestamp int64 `json:"timestamp"`

	// Txn Transaction count for the block as a decimal string.
	Txn string `json:"txn"`
}

// BlocksResponse defines model for BlocksResponse.
type BlocksResponse struct {
	List     []Block `json:"list"`
	Page     int     `json:"page"`
	PageSize int     `json:"pageSize"`

	// Total Total number of blocks matching the filter.
	Total int64 `json:"total"`
}

// Bytes32 A 32-byte value as a `0x`-prefixed hexadecimal string.
type Bytes32 = string

// DailyCommitment defines model for DailyCommitment.
type DailyCommitment struct {
	CommitmentCount int64 `json:"commitment_count"`

	// DailyContractAddress 20-byte Ethereum address as a `0x`-prefixed hex string.
	DailyContractAddress Address `json:"daily_contract_address"`

	// DataType A 32-byte value as a `0x`-prefixed hexadecimal string.
	DataType Bytes32 `json:"data_type"`

	// DayTimestamp UTC day as Unix seconds.
	DayTimestamp   int64 `json:"day_timestamp"`
	DiscoveryBlock int64 `json:"discovery_block"`

	// FactoryAddress 20-byte Ethereum address as a `0x`-prefixed hex string.
	FactoryAddress Address `json:"factory_address"`

	// InstitutionId A 32-byte value as a `0x`-prefixed hexadecimal string.
	InstitutionId Bytes32 `json:"institution_id"`
}

// DailyCommitmentsResponse defines model for DailyCommitmentsResponse.
type DailyCommitmentsResponse struct {
	Limit  int               `json:"limit"`
	List   []DailyCommitment `json:"list"`
	Offset int               `json:"offset"`
}

// ErrorBody defines model for ErrorBody.
type ErrorBody struct {
	// Code Machine-readable error code.
	Code string `json:"code"`

	// Message Human-readable error description.
	Message string `json:"message"`
}

// ErrorResponse defines model for ErrorResponse.
type ErrorResponse struct {
	Error ErrorBody `json:"error"`
}

// TokenTransfer defines model for TokenTransfer.
type TokenTransfer struct {
	BlockNumber int64 `json:"blockNumber"`

	// From Transfer sender (zero address for mints).
	From string `json:"from"`

	// LogIndex Log index within the transaction receipt.
	LogIndex int `json:"logIndex"`

	// To Transfer recipient (zero address for burns).
	To string `json:"to"`

	// TokenAddress 20-byte Ethereum address as a `0x`-prefixed hex string.
	TokenAddress Address `json:"tokenAddress"`

	// TransactionHash 32-byte transaction hash as a `0x`-prefixed hex string.
	TransactionHash TransactionHash `json:"transactionHash"`

	// ValueRaw Transfer amount as a base-10 integer string (token smallest unit).
	ValueRaw string `json:"valueRaw"`
}

// TokenTransfersResponse defines model for TokenTransfersResponse.
type TokenTransfersResponse struct {
	List []TokenTransfer `json:"list"`

	// NextCursor Opaque cursor for the next page. Omitted when there are no more results.
	NextCursor *string `json:"nextCursor,omitempty"`
	PageSize   int     `json:"pageSize"`
}

// Transaction defines model for Transaction.
type Transaction struct {
	BlockNumber int64 `json:"blockNumber"`

	// Data Transaction input data (`0x`-prefixed hex).
	Data string `json:"data"`
	From string `json:"from"`

	// Hash 32-byte transaction hash as a `0x`-prefixed hex string.
	Hash TransactionHash `json:"hash"`

	// Id Placeholder list identifier (always `1` for single-hash lookups).
	Id int64 `json:"id"`

	// IsDataAnchor True when the transaction destination is a registered data-anchor
	// factory or a discovered daily commitment contract.
	IsDataAnchor bool                `json:"isDataAnchor"`
	Metadata     TransactionMetadata `json:"metadata"`

	// Timestamp Block timestamp in milliseconds since Unix epoch.
	Timestamp int64  `json:"timestamp"`
	To        string `json:"to"`
}

// TransactionHash 32-byte transaction hash as a `0x`-prefixed hex string.
type TransactionHash = string

// TransactionMetadata defines model for TransactionMetadata.
type TransactionMetadata struct {
	// FunctionName Resolved method name when known; otherwise `unknown`.
	FunctionName string `json:"functionName"`
}

// GetAddressBalanceParams defines parameters for GetAddressBalance.
type GetAddressBalanceParams struct {
	Block *string `form:"block,omitempty" json:"block,omitempty"`
}

// GetBlocksParams defines parameters for GetBlocks.
type GetBlocksParams struct {
	// Page Page number (1-based).
	Page *int `form:"page,omitempty" json:"page,omitempty"`

	// PageSize Number of blocks per page.
	PageSize *int `form:"pageSize,omitempty" json:"pageSize,omitempty"`

	// MaxBlockNumber Inclusive upper bound on block number (decimal string).
	// When omitted or invalid, the storage layer applies its default upper bound.
	MaxBlockNumber *string `form:"maxBlockNumber,omitempty" json:"maxBlockNumber,omitempty"`

	// OnlyWithTransactions When true, only blocks that contain at least one indexed transaction are returned.
	OnlyWithTransactions *bool `form:"onlyWithTransactions,omitempty" json:"onlyWithTransactions,omitempty"`
}

// GetDailyCommitmentsParams defines parameters for GetDailyCommitments.
type GetDailyCommitmentsParams struct {
	FactoryAddress *Address `form:"factory_address,omitempty" json:"factory_address,omitempty"`

	// DayFrom Inclusive lower Unix-second day timestamp.
	DayFrom *int64 `form:"day_from,omitempty" json:"day_from,omitempty"`

	// DayTo Inclusive upper Unix-second day timestamp.
	DayTo         *int64   `form:"day_to,omitempty" json:"day_to,omitempty"`
	InstitutionId *Bytes32 `form:"institution_id,omitempty" json:"institution_id,omitempty"`
	DataType      *Bytes32 `form:"data_type,omitempty" json:"data_type,omitempty"`
	Limit         *int     `form:"limit,omitempty" json:"limit,omitempty"`
	Offset        *int     `form:"offset,omitempty" json:"offset,omitempty"`
}

// GetTokenTransfersParams defines parameters for GetTokenTransfers.
type GetTokenTransfersParams struct {
	// Cursor Opaque pagination cursor from a previous response `nextCursor`.
	Cursor *string `form:"cursor,omitempty" json:"cursor,omitempty"`

	// PageSize Maximum number of transfers to return.
	PageSize *int `form:"pageSize,omitempty" json:"pageSize,omitempty"`

	// FromBlock Inclusive lower bound block number (decimal string).
	FromBlock *string `form:"fromBlock,omitempty" json:"fromBlock,omitempty"`

	// ToBlock Inclusive upper bound block number (decimal string).
	ToBlock *string `form:"toBlock,omitempty" json:"toBlock,omitempty"`

	// Address Optional filter matching sender or recipient of the transfer.
	Address *Address `form:"address,omitempty" json:"address,omitempty"`
}

// ServerInterface represents all server handlers.
type ServerInterface interface {
	// Get native address balance
	// (GET /api/v1/addresses/{address}/balance)
	GetAddressBalance(w http.ResponseWriter, r *http.Request, address Address, params GetAddressBalanceParams)
	// List indexed blocks
	// (GET /api/v1/blocks)
	GetBlocks(w http.ResponseWriter, r *http.Request, params GetBlocksParams)
	// List indexed daily commitment statistics
	// (GET /api/v1/data-anchor/daily-commitments)
	GetDailyCommitments(w http.ResponseWriter, r *http.Request, params GetDailyCommitmentsParams)
	// List ERC-20 token transfers
	// (GET /api/v1/tokens/{tokenAddress}/transfers)
	GetTokenTransfers(w http.ResponseWriter, r *http.Request, tokenAddress Address, params GetTokenTransfersParams)
	// Get a transaction by hash
	// (GET /api/v1/transactions/{hash})
	GetTransactionByHash(w http.ResponseWriter, r *http.Request, hash TransactionHash)
}

// ServerInterfaceWrapper converts contexts to parameters.
type ServerInterfaceWrapper struct {
	Handler            ServerInterface
	HandlerMiddlewares []MiddlewareFunc
	ErrorHandlerFunc   func(w http.ResponseWriter, r *http.Request, err error)
}

type MiddlewareFunc func(http.Handler) http.Handler

// GetAddressBalance operation middleware
func (siw *ServerInterfaceWrapper) GetAddressBalance(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "address" -------------
	var address Address

	err = runtime.BindStyledParameterWithOptions("simple", "address", r.PathValue("address"), &address, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "address", Err: err})
		return
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params GetAddressBalanceParams

	// ------------- Optional query parameter "block" -------------

	err = runtime.BindQueryParameter("form", true, false, "block", r.URL.Query(), &params.Block)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "block", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetAddressBalance(w, r, address, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetBlocks operation middleware
func (siw *ServerInterfaceWrapper) GetBlocks(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetBlocksParams

	// ------------- Optional query parameter "page" -------------

	err = runtime.BindQueryParameter("form", true, false, "page", r.URL.Query(), &params.Page)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "page", Err: err})
		return
	}

	// ------------- Optional query parameter "pageSize" -------------

	err = runtime.BindQueryParameter("form", true, false, "pageSize", r.URL.Query(), &params.PageSize)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "pageSize", Err: err})
		return
	}

	// ------------- Optional query parameter "maxBlockNumber" -------------

	err = runtime.BindQueryParameter("form", true, false, "maxBlockNumber", r.URL.Query(), &params.MaxBlockNumber)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "maxBlockNumber", Err: err})
		return
	}

	// ------------- Optional query parameter "onlyWithTransactions" -------------

	err = runtime.BindQueryParameter("form", true, false, "onlyWithTransactions", r.URL.Query(), &params.OnlyWithTransactions)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "onlyWithTransactions", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetBlocks(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetDailyCommitments operation middleware
func (siw *ServerInterfaceWrapper) GetDailyCommitments(w http.ResponseWriter, r *http.Request) {

	var err error

	// Parameter object where we will unmarshal all parameters from the context
	var params GetDailyCommitmentsParams

	// ------------- Optional query parameter "factory_address" -------------

	err = runtime.BindQueryParameter("form", true, false, "factory_address", r.URL.Query(), &params.FactoryAddress)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "factory_address", Err: err})
		return
	}

	// ------------- Optional query parameter "day_from" -------------

	err = runtime.BindQueryParameter("form", true, false, "day_from", r.URL.Query(), &params.DayFrom)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "day_from", Err: err})
		return
	}

	// ------------- Optional query parameter "day_to" -------------

	err = runtime.BindQueryParameter("form", true, false, "day_to", r.URL.Query(), &params.DayTo)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "day_to", Err: err})
		return
	}

	// ------------- Optional query parameter "institution_id" -------------

	err = runtime.BindQueryParameter("form", true, false, "institution_id", r.URL.Query(), &params.InstitutionId)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "institution_id", Err: err})
		return
	}

	// ------------- Optional query parameter "data_type" -------------

	err = runtime.BindQueryParameter("form", true, false, "data_type", r.URL.Query(), &params.DataType)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "data_type", Err: err})
		return
	}

	// ------------- Optional query parameter "limit" -------------

	err = runtime.BindQueryParameter("form", true, false, "limit", r.URL.Query(), &params.Limit)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "limit", Err: err})
		return
	}

	// ------------- Optional query parameter "offset" -------------

	err = runtime.BindQueryParameter("form", true, false, "offset", r.URL.Query(), &params.Offset)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "offset", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetDailyCommitments(w, r, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetTokenTransfers operation middleware
func (siw *ServerInterfaceWrapper) GetTokenTransfers(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "tokenAddress" -------------
	var tokenAddress Address

	err = runtime.BindStyledParameterWithOptions("simple", "tokenAddress", r.PathValue("tokenAddress"), &tokenAddress, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "tokenAddress", Err: err})
		return
	}

	// Parameter object where we will unmarshal all parameters from the context
	var params GetTokenTransfersParams

	// ------------- Optional query parameter "cursor" -------------

	err = runtime.BindQueryParameter("form", true, false, "cursor", r.URL.Query(), &params.Cursor)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "cursor", Err: err})
		return
	}

	// ------------- Optional query parameter "pageSize" -------------

	err = runtime.BindQueryParameter("form", true, false, "pageSize", r.URL.Query(), &params.PageSize)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "pageSize", Err: err})
		return
	}

	// ------------- Optional query parameter "fromBlock" -------------

	err = runtime.BindQueryParameter("form", true, false, "fromBlock", r.URL.Query(), &params.FromBlock)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "fromBlock", Err: err})
		return
	}

	// ------------- Optional query parameter "toBlock" -------------

	err = runtime.BindQueryParameter("form", true, false, "toBlock", r.URL.Query(), &params.ToBlock)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "toBlock", Err: err})
		return
	}

	// ------------- Optional query parameter "address" -------------

	err = runtime.BindQueryParameter("form", true, false, "address", r.URL.Query(), &params.Address)
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "address", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetTokenTransfers(w, r, tokenAddress, params)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

// GetTransactionByHash operation middleware
func (siw *ServerInterfaceWrapper) GetTransactionByHash(w http.ResponseWriter, r *http.Request) {

	var err error

	// ------------- Path parameter "hash" -------------
	var hash TransactionHash

	err = runtime.BindStyledParameterWithOptions("simple", "hash", r.PathValue("hash"), &hash, runtime.BindStyledParameterOptions{ParamLocation: runtime.ParamLocationPath, Explode: false, Required: true})
	if err != nil {
		siw.ErrorHandlerFunc(w, r, &InvalidParamFormatError{ParamName: "hash", Err: err})
		return
	}

	handler := http.Handler(http.HandlerFunc(func(w http.ResponseWriter, r *http.Request) {
		siw.Handler.GetTransactionByHash(w, r, hash)
	}))

	for _, middleware := range siw.HandlerMiddlewares {
		handler = middleware(handler)
	}

	handler.ServeHTTP(w, r)
}

type UnescapedCookieParamError struct {
	ParamName string
	Err       error
}

func (e *UnescapedCookieParamError) Error() string {
	return fmt.Sprintf("error unescaping cookie parameter '%s'", e.ParamName)
}

func (e *UnescapedCookieParamError) Unwrap() error {
	return e.Err
}

type UnmarshalingParamError struct {
	ParamName string
	Err       error
}

func (e *UnmarshalingParamError) Error() string {
	return fmt.Sprintf("Error unmarshaling parameter %s as JSON: %s", e.ParamName, e.Err.Error())
}

func (e *UnmarshalingParamError) Unwrap() error {
	return e.Err
}

type RequiredParamError struct {
	ParamName string
}

func (e *RequiredParamError) Error() string {
	return fmt.Sprintf("Query argument %s is required, but not found", e.ParamName)
}

type RequiredHeaderError struct {
	ParamName string
	Err       error
}

func (e *RequiredHeaderError) Error() string {
	return fmt.Sprintf("Header parameter %s is required, but not found", e.ParamName)
}

func (e *RequiredHeaderError) Unwrap() error {
	return e.Err
}

type InvalidParamFormatError struct {
	ParamName string
	Err       error
}

func (e *InvalidParamFormatError) Error() string {
	return fmt.Sprintf("Invalid format for parameter %s: %s", e.ParamName, e.Err.Error())
}

func (e *InvalidParamFormatError) Unwrap() error {
	return e.Err
}

type TooManyValuesForParamError struct {
	ParamName string
	Count     int
}

func (e *TooManyValuesForParamError) Error() string {
	return fmt.Sprintf("Expected one value for %s, got %d", e.ParamName, e.Count)
}

// Handler creates http.Handler with routing matching OpenAPI spec.
func Handler(si ServerInterface) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{})
}

// ServeMux is an abstraction of http.ServeMux.
type ServeMux interface {
	HandleFunc(pattern string, handler func(http.ResponseWriter, *http.Request))
	ServeHTTP(w http.ResponseWriter, r *http.Request)
}

type StdHTTPServerOptions struct {
	BaseURL          string
	BaseRouter       ServeMux
	Middlewares      []MiddlewareFunc
	ErrorHandlerFunc func(w http.ResponseWriter, r *http.Request, err error)
}

// HandlerFromMux creates http.Handler with routing matching OpenAPI spec based on the provided mux.
func HandlerFromMux(si ServerInterface, m ServeMux) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseRouter: m,
	})
}

func HandlerFromMuxWithBaseURL(si ServerInterface, m ServeMux, baseURL string) http.Handler {
	return HandlerWithOptions(si, StdHTTPServerOptions{
		BaseURL:    baseURL,
		BaseRouter: m,
	})
}

// HandlerWithOptions creates http.Handler with additional options
func HandlerWithOptions(si ServerInterface, options StdHTTPServerOptions) http.Handler {
	m := options.BaseRouter

	if m == nil {
		m = http.NewServeMux()
	}
	if options.ErrorHandlerFunc == nil {
		options.ErrorHandlerFunc = func(w http.ResponseWriter, r *http.Request, err error) {
			http.Error(w, err.Error(), http.StatusBadRequest)
		}
	}

	wrapper := ServerInterfaceWrapper{
		Handler:            si,
		HandlerMiddlewares: options.Middlewares,
		ErrorHandlerFunc:   options.ErrorHandlerFunc,
	}

	m.HandleFunc("GET "+options.BaseURL+"/api/v1/addresses/{address}/balance", wrapper.GetAddressBalance)
	m.HandleFunc("GET "+options.BaseURL+"/api/v1/blocks", wrapper.GetBlocks)
	m.HandleFunc("GET "+options.BaseURL+"/api/v1/data-anchor/daily-commitments", wrapper.GetDailyCommitments)
	m.HandleFunc("GET "+options.BaseURL+"/api/v1/tokens/{tokenAddress}/transfers", wrapper.GetTokenTransfers)
	m.HandleFunc("GET "+options.BaseURL+"/api/v1/transactions/{hash}", wrapper.GetTransactionByHash)

	return m
}